MGS-5226 [#imx-1746] armnn-0001-NNRT-10-unittest-failed-at-DepthWiseConvolution2d-wi
[NNRT-10]unittest failed at DepthWiseConvolution2d with
 multiplier >= 2

Reviewed-on: http://192.168.32.113/review/4921
Tested-by: xiang.zhang <Xiang.Zhang@verisilicon.com>

+++ nn_runtime/armnn_backend_vsi_npu/workloads/FakeBiasSelector.hpp
@@ -46,7 +46,7 @@ struct FakeBiasTypeSelector<DataType::Float16> {
 
 template <>
 struct FakeBiasTypeSelector<DataType::QuantisedAsymm8> {
-    using type = uint8_t;
+    using type = int32_t;
     static constexpr armnn::DataType value = DataType::Signed32;
 };
